Move recovery from /etc/functions to /bin
TPM2 cannot extend shell functions, only full binaries.  Thus move
recovery to its own file in /bin as it's the most common thing to be
extended.  An alternative is to use the tpm2 pcrextend with a
precomputed pcr digest, however that will not allow the use of the
generic tpm_extend function.

Signed-off-by: Cody Ho <codyho@stanford.edu>